Viewer: Extend NoReconnect to initial conn./auth.
Offer to reconnect if the initial connection or authentication fails.
Note that the viewer will only offer to reconnect during the initial
connection/authentication phase if a WarningException is thrown.
Effectively, that means that it will offer to reconnect due to a normal
connection failure (for instance, specifying the wrong hostname or
display number)	or a normal authentication failure (for instance,
mistyping the password), but it will not offer to reconnect if the
maximum number of consecutive VNC password or OTP authentication
failures has been exceeded.  Furthermore, if an authentication failure
occurs, the viewer will return to the authentication dialog, but if any
other failure occurs, it will return to the "New TurboVNC Connection"
dialog (thus allowing the user to adjust the hostname, display number,
gateway, etc.)

This commit also modifies CConn to ensure that session manager/tunneling
errors due to an unknown host are thrown as WarningExceptions, thus
causing the viewer to offer to reconnect.

Closes #92