RPM: Exclude custom JRE from auto. dep. processing
This prevents the RPM packages from exporting any internal dependencies
from the custom JRE that are resolved by same.  This fixes an issue
whereby TurboVNC RPMs built on RHEL 7+ (and derivatives) could not be
installed on RHEL 9 (and derivatives.)  Since the issue did not affect
our official RPMs, the fix is not documented in the change log.