Java: Properly document -config option
A .vnc connection info file can only be loaded by passing -config <file>
or <file> to the Java TurboVNC Viewer.  Documenting that behavior by way
of a fake "Config" parameter was confusing, because it implied that a
connection info file could be loaded with Config=<file> or
--config <file>, etc.  It makes more sense to document -config as an
option, since it works similarly to -loglevel.  Also, the fact that the
viewer treats the VNC server argument as a connection info file if it
ends in ".vnc" was never documented.

NOTE: Technically speaking, a connection info file does not need a .vnc
extension in order to be passed to -config or dragged & dropped.  Chalk
that up to a white lie told in the name of simplifying the
documentation.