Un*x vncviewer: Don't pass -d32 to 32-bit JRE
4baa42918508c1b77edba2e8675707046b8609d1 removed the -d64 option that
the Un*x vncviewer script traditionally passed to the java executable,
because Java 9+ no longer supports that option.  However, the -d32
option was retained when using a 32-bit JRE.  That decision was probably
based on the false assumption that, because Oracle doesn't provide a
32-bit Java 9+ JRE for Un*x, one can only use Java 8- with 32-bit Un*x
systems.  However, there are some 32-bit Linux distributions out there
that compile their own 32-bit OpenJDK 9+ implementations, and the use of
-d32 breaks those implementations.