Viewer: Remove *Ident sec. types + ClientRedirect
The *Ident security types, which pass only a username to the server,
were originally implemented by Brian Hinz in order to support a specific
portal solution, but the security types are not documented
(https://github.com/rfbproto/rfbproto/blob/master/rfbproto.rst), nor
have they presumably been incorporated into the VeNCrypt specification.
To the best of my knowledge, they are unused and untested in the
TurboVNC Viewer.  The ClientRedirect pseudo-encoding, which I believe
was part of the same portal solution, is registered in the community RFB
spec but not documented.  To the best of my knowledge, it is also unused
and untested in the TurboVNC Viewer, and it conflicts with a
pseudo-encoding implemented by RealVNC.  (See
20d1d4ae14a86beb11d1498031490046e997a182.)