Mac: Add parameter to disable menu accelerators
Referring to #345, some users need to use Super/Meta key combinations
with Emacs, and since the Command key on macOS maps to the Super/Meta
key on remote Un*x systems, some of those key combinations are hijacked
by the Mac menu accelerators before they can be sent to the VNC server.
This commit introduces a new parameter (MacHotkeys) that can be used to
disable the Mac menu accelerators.  This parameter is currently
undocumented because it is an incomplete solution to a
rarely-encountered problem.  (The complete solution would be to
implement keyboard grabbing on macOS.  Refer to #162.)