Windows: Use UTF-8 for clipboard String/byte conv.
Java uses UTF-8 as the default character set on macOS and Linux, but on
Windows, it uses windows-1252.  Thus, the TurboVNC Viewer needs to
explicitly specify the UTF-8 character set when converting the clipboard
String to a byte array prior to sending and when converting the
clipboard byte array to a String after receiving.  This issue could also
be worked around by setting the file.encoding system property to UTF-8
(e.g. by setting JAVA_TOOL_OPTIONS=-Dfile.encoding=UTF-8 in the
environment), which changes the default character set.

Fixes #375