Server: Remove in-tree X.org support libraries
All supported TurboVNC host platforms now include sufficiently recent
versions of these, so it is no longer necessary to provide in-tree
versions.  Effectively, this means that TVNC_SYSTEMX11 and
TVNC_SYSTEMLIBS are now always on.  This commit introduces a new CMake
variable (TVNC_INTELZLIB) to toggle the use of the in-tree Intel zlib
implementation (as opposed to the system's zlib implementation) on x86
platforms.

This commit also removes the TVNC_GLX CMake variable, because an X
server without a GLX extension is pretty useless these days.

This commit also sets TVNC_BUILDSERVER to 0 by default on Solaris.
Solaris/x86 11.4 (at least the GA version-- haven't tested the latest
SRU) does not have new enough X.org support libraries to build the
TurboVNC Server (at least not without a great deal of effort.)  We have
not claimed to officially support Solaris for quite some time, and the
only reason it was still tested with the TurboVNC Server was the fact
that nVidia continues to support it.  This seems like a good time to
stop testing it.