Build: Show TVNC_GLX, TVNC_INCLUDEJRE vals as 0/1
This is purely cosmetic.  For CMake boolean variables that are used in
packaging scripts, we call boolean_number(<variable> PARENT_SCOPE) to
set the variable's value to 0 or 1 in the parent scope, which is what
the packaging scripts expect.  However, we also need to call
boolean_number(<variable>) to set the variable's value to 0 or 1 in the
current scope.  This makes the output of report_option() consistent for
all boolean variables, regardless of whether they were specified as 1/0,
TRUE/FALSE, ON/OFF, YES/NO, or Y/N.  It also directs users toward the
most intuitive and linguistically-neutral way to set or change a boolean
variable's value.