Viewer: Allow QEMU key ext without LED State exts
The QEMU and VMware LED State RFB extensions synchronize the lock key
states between the client and the VNC server, but those extensions
aren't strictly necessary when using the QEMU Extended Key Event
RFB extension.  More specifically, wayvnc/neatvnc supports the QEMU
Extended Key Event extension but not the LED State extensions, so the
TurboVNC Viewer was silently falling back and sending regular RFB key
events to wayvnc.

This commit causes the TurboVNC Viewer to emulate the behavior of the
TigerVNC Viewer, which allows the use of the QEMU Extended Key Event
extension even if no LED State extension is available.  This is a
non-issue when using the TurboVNC Viewer with the TurboVNC Server, which
supports the LED State extensions.  With wayvnc and other VNC servers
that support the QEMU Extended Key Event extension but not the LED State
extensions, it means that the lock key state on the client will lose
synchronization with the lock key state on the remote desktop if a lock
key is pressed outside of the TurboVNC Viewer window.  (The same is true
when using the TigerVNC Viewer with wayvnc.)

Fixes #377