vncserver: Check a display #'s abstract socket
... before assuming that the display number is available.

Newer versions of GDM use Display :0 for the greeter and Display :1 for
the session if WaylandEnable=false.  However, X11 lock files are no
longer used, and the session's X server still functions if its Unix
domain socket (/tmp/.X11-unix/X1) is deleted.  Thus, if a user logged in
locally, accidentally deleted /tmp/.X11-unix/X1, then tried to start a
TurboVNC session, vncserver would attempt to use Display :1, and the
window manager in the TurboVNC session would subsequently fail to start.

Based on:
https://github.com/algorythmic/turbovnc/commit/1028467a2cc29f4e65589bee15ff649278a51347

Note that the aforementioned commit attempts to actually open the
Unix domain socket rather than checking for the existence of the socket
file.  However, I could find no cases in which the ability to open a
Unix domain socket was not determined by the existence of the socket
file.