vncserver: Silence wrng. when listing UDS sessions
The List() subroutine iterates through the .pid files in the TurboVNC
user directory, calling CheckDisplayNumber() for each corresponding
display number to verify that the corresponding TurboVNC session is
still running.  However, we need to silence the warnings in
CheckDisplayNumber() when we use it for that purpose.  Otherwise, if the
session is listening on a Unix domain socket, CheckDisplayNumber() will
fall through the TCP port checks and print a warning that the abstract
socket corresponding to the session's display number is in use.  That
warning is only relevant when starting a new session.