Viewer: Send horizontal scroll wheel events
On macOS and Windows, these are communicated to applications as vertical
scroll wheel events with the Shift key held down.  (The operating system
acts as if the Shift key is held down, but no Shift key press/release
events are actually fired.)  We use RFB Buttons 6 and 7 for,
respectively, left and right scroll wheel events, since X11 Buttons 6
and 7 are used for the same purpose.  (Thus, implementing this feature
in the TurboVNC Server simply required incrementing the loop limit in
PtrAddEvent().)  Java doesn't need to use Java Buttons 4 and 5 for
vertical scroll wheel events, since those events have a separate
subclass.  Thus, X11 Buttons 6 and 7 map to Java Buttons 4 and 5, and we
simply map Java Buttons 4 and 5 to RFB Buttons 6 and 7 in order to
implement horizontal scrolling for X11 clients.

Server and macOS/Windows viewer implementation based on:
https://github.com/rsparlin/turbovnc/commit/77e266d6deaa1df6b537cfd3599c445e90a378db

Closes #416