Server: Allow max Tight subrect size to be tweaked
Referring to
https://groups.google.com/g/turbovnc-devel/c/LDSsexdiDXs/m/XC3LxDTpAQAJ
this can be useful with viewers that have a high amount of computational
overhead (e.g. noVNC) running in a loopback environment (e.g. to connect
to a virtual machine.)

This feature is currently undocumented because it benefits only esoteric
use cases, and the full scope of its trade-offs is not well understood.

A better long-term solution is probably pure JPEG encoding (#376), which
noVNC supports, but that would require changes to the TurboJPEG API
(libjpeg-turbo/libjpeg-turbo#777) that haven't landed yet.