JSch: Fix detection of supported RSA sig schemes
KeyExchange.guess() only returns the first server/client match for each
category, and the client algorithms are taken from the HostKeyAlgorithms
OpenSSH config file keyword rather than the PubkeyAcceptedAlgorithms
keyword.  Thus, fd34df2ca25f99288f0569511b68f107c289e7b1 effectively
made it so that an RSA signature scheme could only be used if it was
the first server-supported algorithm listed with the HostKeyAlgorithms
keyword.

Instead, set Session.supportedRSAMethods to the list of RSA signature
schemes that the server supports, and attempt to use the first one of
those algorithms that is specified with PubkeyAcceptedAlgorithms.  This
fulfills the intent of fd34df2ca25f99288f0569511b68f107c289e7b1 and
emulates the behavior of OpenSSH.