Viewer: Fix accepting clipboard data from QEMU's VNC server
The ServerCutText sent by the QEMU's VNC server contains a complete
zlib stream with a final block at the end. Thus, when inflate() is called
to decompress the zlib stream, inflate() will return JZlib.Z_STREAM_END
when the end of stream is reached.

In this case, it is not an error, and we should not throw an exception.

Signed-off-by: Yan-Jie Wang <yanjiewtw@gmail.com>