Fix build if TVNC_SYSTEMLIBS && !TVNC_SYSTEMX11
If TVNC_SYSTEMLIBS and TVNC_SYSTEMX11 are both set, then the build
system will look for the XFont2 development package, which depends on
the FreeType, bzip2, and zlib development packages.  However, if
TVNC_SYSTEMLIBS is set and TVNC_SYSTEMX11 is unset, then the build
system needs to explicitly look for those packages and fail if one of
them isn't available.  The easiest way to do that is with
find_package(... REQUIRED).

Closes #431