Viewer: Fix SSH port configuration hierarchy
dda0283384b82c62b930c6516f6c19a59e8a1a46 intended to implement the
following configuration hierarchy for the SSH port:

1. The SSHPort parameter (specified on the command line or in a
   connection info file)
2. The OpenSSH config file (specified in the SSHConfig parameter, which
   defaults to ~/.ssh/config)
3. The default value of the SSHPort parameter (specified in
   ~/.vnc/default.turbovnc, otherwise 22)

Due to an oversight, however, the initial implementation required the
SSH port to be specified either in the SSHPort parameter or an OpenSSH
config file, or an IllegalArgumentException occurred (as a result of -1
being passed to the java.net.Socket constructor.)
053e754cb720bd9d4e07e0e8ca43222f0c312afc attempted to address that issue
by modifying com.jcraft.jsch.Session.applyConfig(), but that method is
only called if an OpenSSH config file exists or is specified.  Otherwise
an IllegalArgumentException still occurred.  Furthermore, the value of
SSHPort in ~/.vnc/default.turbovnc was ignored.  This commit addresses
both issues.