JSch: Fix whitespace issues
- Remove trailing spaces and new lines.
- Convert tabs to spaces.

There will be no new releases of JCraft JSch from which to pull, so we
no longer need to maintain the whitespace issues in that code.
Eventually I would like to format our JSch fork similarly to the rest of
the TurboVNC Viewer code base, but that's a problem for another day.  At
least now I don't have to reconfigure my text editor every time I change
something in JSch.