Viewer: Add parameter to disable hotkeys
This is a companion to the NoMacHotkeys parameter, to address rare cases
in which users may want to pass all Ctrl-Alt-Shift key sequences to the
VNC server.  (All Ctrl-Alt-Shift hotkeys have an equivalent F8 menu
hotkey, so one could use, for instance, {menu key} + F instead of
Ctrl-Alt-Shift-F in order to toggle full-screen mode.)

Also change "key combination" to "key sequence" in the NoMacHotkeys
parameter documentation, for consistency.