Viewer: Use separate parameter for jump host
The jump host feature is an improvement upon the traditional Via
feature, since it eliminates the need to open RFB ports in the VNC
host's firewall, it ensures that the RFB connection is encrypted on the
server-area network, and it allows the TurboVNC Session Manager to be
used with TurboVNC hosts that are behind an SSH gateway.  Thus, I
initially thought it best to repurpose the Via parameter to specify a
jump host and to hide the traditional Via functionality behind a Java
system property (turbovnc.viajump.)  This was effectively a deprecation
model for the traditional Via feature and effectively made the jump host
feature opt-out instead of opt-in.

However, because Via can also be used to specify an UltraVNC Repeater
gateway, repurposing the parameter made it impossible to specify
different SSH ports for the jump host and VNC host without using an
OpenSSH config file.  I could have hidden the UltraVNC Repeater gateway
feature behind turbovnc.viajump as well, but that would have created a
backward compatibility issue.  (For instance, if saved options from a
previous version of TurboVNC specified Via={gateway_host}:1, then
TurboVNC 3.2 would have interpreted 1 as an SSH port instead of an
UltraVNC Repeater display number.)  I could have also added a new
parameter (ViaSSHPort, for example) for specifying the jump host's SSH
port, but ultimately it was less confusing to just introduce a dedicated
parameter for specifying a jump host and to leave Via alone.  The new
Jump parameter behaves identically to ssh -J, except that it does not
allow multiple comma-separated hops to be specified.

Closes #440