Viewer: Fix rare deadlock w/ auto desktop resizing
(regression introduced by 01549361cb5f85b93403feb36df8ab5694a6f962)

If automatic desktop resizing is enabled and the TurboVNC Viewer is in
windowed mode, recreating the viewer window in the body of
CConn.resizeFramebuffer() sometimes causes the window's focusLost()
method to be invoked.  Under rare circumstances, the method can be
invoked in the EDT while the RFB thread is still in the body of
CConn.resizeFramebuffer().  Because of
01549361cb5f85b93403feb36df8ab5694a6f962, CConnection.processMsg()
(in the RFB thread) obtains a monitor lock on the CConnection/CConn
instance, and the lock is still owned by the RFB thread when
CConn.resizeFramebuffer() is called from CConnection.processMsg().  When
the viewer window's focusLost() method attempted to call
CConn.releasePressedKeys() in the EDT, CConn.releasePressedKeys() failed
to obtain a monitor lock on the CConnection/CConn instance, and a
deadlock occurred.

It isn't actually necessary for CConn.releasePressedKeys() to obtain a
monitor lock.  That method is only ever executed in the EDT, and the
resources it accesses are only ever accessed in the EDT.  The monitor
lock was in the original code base that the TurboVNC Viewer inherited
from TigerVNC, but TigerVNC no longer has it.  (The 'synchronized'
keyword was silently removed in
TigerVNC/tigervnc@e5b4003bfb24f740e0d6ea2c82422d1dacf7b927.)

Fixes #447