Viewer: Work around JFrame.getInsets()/GNOME bug
On recent versions of GNOME (v43 or v44 and later), and only on recent
versions of GNOME, our hackish method of obtaining the global JFrame
insets for X windows returns { 0, 0, 0, 0 }.  This seems to be due to a
bug in the sun.awt.X11.XDecoratedPeer and sun.awt.X11.XWM classes that
causes the insets to be set from an invalid X11 property after they have
already been set from the _NET_FRAME_EXTENTS property.  Fortunately, on
the affected window managers (and only on the affected window managers),
the cross-platform method of obtaining the global JFrame insets works.