Build Intel JavaAppLauncher w/ Xcode 9.4/10.9 SDK
This is necessary in order to notarize the application and also to build
it on modern macOS platforms.  Since the Adoptium build of OpenJDK 21
requires macOS 10.12 or later, this does not affect the TurboVNC
Viewer's system requirements.

I also took the opportunity to revisit #350, including poring over the
macOS 10.14 SDK change log and trying to manually reverse any changes to
the Cocoa application defaults that landed in that SDK.  The cause of
the issue unfortunately remains a mystery, which means that it cannot
currently be solved or worked around on Apple Silicon platforms.  In the
long term, it may be necessary to work around it by implementing a full-
screen mode that doesn't use Spaces, a la TigerVNC or VLC, and either
abandoning the Spaces-based full-screen mode or retaining it as a
non-default option that doesn't support multi-screen spanning.