Viewer: Don't require Tunnel w/ sess mgr & ExtSSH
Using an external SSH client with the TurboVNC Session Manager is not
officially documented.  However, it works with the caveat that the
external SSH client is only used for the final VNC connection.  (The
built-in SSH client is still used for session management.)
ae1c3fafe02982a871c1d30eb21c8771546cfd56 introduced a regression
whereby, when attempting to use ExtSSH with the TurboVNC Session
Manager, the viewer used the Via command-line template by default
instead of the Tunnel command-line template, which failed because no
gateway host had been specified.