Viewer/TVNC_SINGLESCREEN: Fix dflt FS win size/pos
More specifically, create a new ScreenSet instance in
CConn.computeScreenLayout() if TVNC_SINGLESCREEN/turbovnc.singlescreen
is enabled.  Otherwise the new layout computed in the body of
CConn.sizeWindow() will always be equal to cp.screenLayout, and the new
layout will never be sent to the server.  Effectively, if the
full-screen viewer window size/position was not the default, then
resetting it to the default did not work properly.  (If the viewer
window did not fill the screen, then its position changed, but its size
and the remote desktop size did not change.)

(oversight from 8c275dac536f117e715fc9d72290acb492b7f2db)