Viewer: Fix null ptr exception in SessMgr dialog
If there are no TurboVNC sessions running under the user's account on
the host, then 'sessions' will be null in the SessionManagerDialog
constructor.  Thus, we cannot check the session limit to see whether
the "New session" button should be disabled.  (If there are no running
sessions, then the button shouldn't be disabled anyhow.)