Server: Incr UTF-8 CB xfer buf size to handle CL 0
If zlib compression level 0 is used, then the size of the "compressed"
zlib buffer will exceed the size of the uncompressed buffer.
compressBound() accounts for that overhead, but SendClipboardProvide()
erroneously assumed that the overhead would be incurred only once
rather than for each time the stream is flushed.  Thus, using zlib
compression level 0 frequently caused the zlib stream of UTF-8 clipboard
transfers to be truncated, leading to fatal viewer errors.

Fixes #472